refactored duplicated code into Navbar component
